+++ perfcake/src/main/java/org/perfcake/message/generator/RampUpDownGenerator.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -27,6 +27,7 @@ import org.perfcake.message.sender.MessageSenderManager;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Semaphore;
@@ -129,6 +130,7 @@ public class RampUpDownGenerator extends DefaultMessageGenerator {
 
    @Override
    public void init(final MessageSenderManager messageSenderManager, final List<MessageTemplate> messageStore) throws PerfCakeException {
+      messageSenderManager.setSenderPoolSize(Arrays.asList(preThreadCount, mainThreadCount, postThreadCount, super.getThreads()).stream().reduce(Integer::max).get());
       super.init(messageSenderManager, messageStore);
       if (preThreadCount <= 0) {
          preThreadCount = super.getThreads();