+++ perfcake/src/main/java/org/perfcake/message/sequence/PrimitiveNumberSequence.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -21,28 +21,24 @@ package org.perfcake.message.sequence;
 
 import org.perfcake.PerfCakeException;
 
+import java.util.concurrent.atomic.AtomicLong;
+
 /**
- * Represents an automatically generated sequence of values.
- * The resulting values can be used in the message body.
+ * Just an ever increasing number sequence. No tweaking available. Non-blocking.
  *
  * @author <a href="mailto:marvenec@gmail.com">Martin Večeřa</a>
  */
-public interface Sequence {
+public class PrimitiveNumberSequence implements Sequence {
 
-   /**
-    * Gets the next value in this sequence. Must be thread safe. It is called once per message.
-    *
-    * @return The next value in this sequence.
-    */
-   String getNext();
+   private AtomicLong number = new AtomicLong(0);
 
-   /**
-    * Resets the sequence.
-    * This method is called at the very beginning, so it can be used to perform any initialization steps as well.
-    *
-    * @throws PerfCakeException
-    *       When it was not possible to initialize the sequence to its original state.
-    */
-   default void reset() throws PerfCakeException {
+   @Override
+   public String getNext() {
+      return String.valueOf(number.getAndIncrement());
+   }
+
+   @Override
+   public void reset() throws PerfCakeException {
+      number.set(0);
    }
 }