+++ perfcake/src/main/java/org/perfcake/reporting/reporters/ResponseTimeHistogramReporter.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -112,7 +112,7 @@ public class ResponseTimeHistogramReporter extends AbstractReporter {
    public void publishResult(final PeriodType periodType, final Destination destination) throws ReportingException {
       final Measurement m = newMeasurement();
 
-      DoublePercentileIterator pi = null;
+      DoublePercentileIterator pi;
 
       switch (correctionMode) {
          case AUTO: