+++ perfcake/src/main/java/org/perfcake/util/Utils.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -114,7 +114,7 @@ public class Utils {
 
       matcher.reset();
       while (matcher.find()) {
-         String pValue = null;
+         String pValue;
          final String pName = matcher.group(2);
          String defaultValue = null;
          if (matcher.groupCount() == 3 && matcher.group(3) != null) {