+++ perfcake/src/test/java/org/perfcake/message/sequence/PrimitiveNumberSequenceTest.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,30 +19,21 @@
  */
 package org.perfcake.message.sequence;
 
-import org.perfcake.PerfCakeException;
+import org.testng.Assert;
+import org.testng.annotations.Test;
 
 /**
- * Represents an automatically generated sequence of values.
- * The resulting values can be used in the message body.
- *
  * @author <a href="mailto:marvenec@gmail.com">Martin Večeřa</a>
  */
-public interface Sequence {
+public class PrimitiveNumberSequenceTest {
 
-   /**
-    * Gets the next value in this sequence. Must be thread safe. It is called once per message.
-    *
-    * @return The next value in this sequence.
-    */
-   String getNext();
+   @Test
+   public void testPrimitiveNumberSequence() {
+      final Sequence s = new PrimitiveNumberSequence();
 
-   /**
-    * Resets the sequence.
-    * This method is called at the very beginning, so it can be used to perform any initialization steps as well.
-    *
-    * @throws PerfCakeException
-    *       When it was not possible to initialize the sequence to its original state.
-    */
-   default void reset() throws PerfCakeException {
+      Assert.assertEquals("0", s.getNext());
+      Assert.assertEquals("1", s.getNext());
+      Assert.assertEquals("2", s.getNext());
    }
+
 }